+++ test/unit/exports.ts
@@ -31,6 +31,7 @@
 import * as index from '../../src/index';
 import * as fallback from '../../src/fallback';
 import * as assert from 'assert';
+import {describe, it} from 'mocha';
 
 const version = require('../../../package.json').version;