+++ test/unit/fallbackError.ts
@@ -30,6 +30,7 @@
  */
 
 import * as assert from 'assert';
+import {describe, it} from 'mocha';
 import * as fs from 'fs';
 import * as path from 'path';
 import {FallbackErrorDecoder} from '../../src/fallbackError';